[#82] - fix: 투표 결과 0%로 나오던 문제 수정

+++ domain/src/main/java/com/jwd/lunchvote/domain/usecase/CalculateSecondVote.kt
@@ -28,7 +28,7 @@ class CalculateSecondVote @Inject constructor(
     val voteResult = SecondVoteResult(
       loungeId = loungeId,
       foodId = electedFoodIds.random(),
-      voteRatio = (maxScore/secondBallotList.size).toFloat()
+      voteRatio = maxScore.toFloat()/secondBallotList.size.toFloat()
     )
     voteResultRepository.saveSecondVoteResult(voteResult)
   }
